web/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.1.3
python-dotenv==1.0.0
geopy==2.4.1 